refactor: lottoNumberParser domian 디렉토리로 이동
